replaces the lodash.template function with a roll-our-own global replace
lodash.template uses a construct that is deemed unsafe by at least atom - probably because it in principle is

There are two ways to circumvate this - use atom/loophole to allow _.template as a safe option.

Use an alternative.

The second one is the simplest thing that works, hence I chose that.