feature(*): introduce a 'module' index (#51)
Adds ES2015 modules to the distribution:

ES2015 modules in dist/es2015
a module key in package.json to indicate to webpack || rollup where to find them
a sideEffects key (with value false) to indicate there's no side-effects in this package.
Also reorganises the dist folder a bit - it now has separate folders for the cjs distribution (the default one - it's where main in package.json points to), es2015 (see above) and bundle - which is a webpack'ed, minimised version of the css distribution.

Why?

es2015 modules have some benefits in use over cjs - it a.o. enables build tools to weed out unused code ('tree shaking').